Disable inline editor buttons if view in read-only mode jmix-framework/jmix_4329
